fdsef

+++ projetFilRouge/src/main/java/com/inetum/pfr/projetFilRouge/services/AbstractGenericService.java
@@ -12,6 +12,8 @@ public abstract class AbstractGenericService<E, ID, DTO> implements GenericServi
 	public abstract CrudRepository<E, ID> getDao();
 	public abstract Class<DTO> getDtoClass();
 	
+	// Méthodes renvoyant des <Entity>
+	
 	@Override
 	public E searchById(ID id) throws NotFoundException {
 		E e = getDao().findById(id).orElse(null);
@@ -29,7 +31,13 @@ public abstract class AbstractGenericService<E, ID, DTO> implements GenericServi
 
 	@Override
 	public void removeById(ID id) {
-		getDao().deleteById(id);
+		E e = getDao().findById(id).orElse(null);
+		if(e != null) {
+			getDao().deleteById(id);;
+		} else {
+			throw new NotFoundException("Suppression annulée, aucune correspondance trouvé pour l'id : " + id);
+		}
+		
 	}
 
 	@Override
@@ -41,7 +49,19 @@ public abstract class AbstractGenericService<E, ID, DTO> implements GenericServi
 	public List<E> searchAll() {
 		return (List<E>) getDao().findAll();
 	}
+	
+	// Méthodes renvoyant des <DTO>
 
-
+	@Override
+	public DTO searchDtoById(ID id) throws NotFoundException {
+		E e = this.searchById(id);
+		return GenericConverter.map(e, getDtoClass());
 	}
+
+	@Override
+	public List<DTO> searchAllDto() {
+		List<E> le = (List<E>) getDao().findAll();
+		return GenericConverter.map(le, getDtoClass());
+	}
+}
